Write orders to Firebase
Place orders into Firebase, saving the user_name as the key for
that order. The exact break down of the command format can be
found in commands.js. Still need a way to associate user_name with
name to add onto order on Seamless. Will probably need to make a
separate JSON object on Firebase to hold all of those.