Added checking for valid items
Alfred will now check against the list of valid items in littleDelhi.json
to ensure entries are valid. This is required becuse casper is very dumb!

A tree of little delhi items, the keys are the names that people should
use when placing their order. If we find that people most often use a
different handle for an item, an extra branch can be added easily.