Authorization and clean up
The biggest thing is that now Alfred will only listen to messages
containing the proper token, which #ot-lil-delhi sends. Also, authorization
has been added for the actions that can be done on Firebase, using the
secret for my firebase DB.

And I'm still messing around with error handling, I may change this more.