Changed data structure, added phone number entry
Changed the way data is stored on Firebase to better isolate data
to respective areas. Not sure if this is the proper way to do this
with a key-value store, but it seems cleaner to me than storing name,
number, and items on each order. All the information pertaining to
the user is now stored under 'users'.

Also added the ability to receive phone numbers. Alfred will run a
RNG when Casper asks it for the orders to decide whose phone number
will be entered into Seamless and announced on slack.