Account for incorrect spice options
Only some items have the ability to set a spice level. We will now check
if an item with a user specified spice level is actually an item that you
can set a spice on, otherwise it will notify the user that the item doesn't
have a spice level, remove the spice, and add it to the order.