Add time saved to stats
I did a very dirty and quick estimate to see how long the ordering process
actually takes. For me, it took about 30 seconds for the normal stuff like
logging in and navigating to LD, and each item took about 5 seconds. This is
a lower bound estimate, I'm sure it would usually take twice as long.

I've purposely left out the personal time stat because it's not really an
impressive number to look at, it would be on the order of minutes per person.