Add scrape script
It's not at all feasible for me to manually type up each item from each restaurant without
going insane. I wrote this scraper to do that for me, which almost drove me insane anyhow
because I had to learn how to Regex. But since I've done so, I was able to just run the
script and get all the information I need from Seamless.

It's highly unlikely that Seamless will change anything about these pages, so I doubt I'll
ever need to run the script again, but I'm adding it to the code anyhow, to document my
struggles :( .

This script was largely copied from the existing casper.js script, which I'll probably end
up heavily modifying soon. Most of the interesting stuff I've done for this script are in
the comments, so check them out if you wanna descend into the madness.