Get all options
Previously I'd only scrape the radio options, but to be truly useful (especially
for pizza ordering), you need to have all the options, including the checkboxes.
I've also added a category option so we can list the options with more data, and
a boolean to show whether the category is a single choice one.