Use Watson to parse orders
The Watson library takes the response from Watson, and parses it to get the
restaurant, menu items, and options that the user wants. There is a lot going
on there, and it is extensively commented, so check out the comments if you wanna
know how it works.

This commit also sets up the basic ground work for the beta of alfie3, the
unnecessary commands have been removed from the app. The commands file has been
renamed and significantly trimmed down. I've also added some formatting for
slack, so it shows up as richly formatted.