Adjust Watson entities
- an END_TOKEN (*) has been added to all values. This is because some names on seamless
are followed by a space. When these names are uploaded to watson, it truncates the space,
so when a synonym is recognized, the value without the space is what is returned. Because
CasperJS is very exact on what constitutes a click "match", this causes issues. To remedy
this problem, I'm appending a token to the end of the value, which I will strip out upon
receiving the value back from Watson.

- Removed the two words regex for options and menu items because it was registering for
phrases that weren't even meant to be orders. I think I'll add a separate file that we
can add one off abbreviations for certain items instead of trying to force that with this
regex.

- Filter out short synonyms, we want at least 3 characters, to avoid overlap with random
phrases.